requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
aiohttp
python-multipart
pinecone[grpc]
google-genai
